Synchronize steps into stage status of PipelineRun (#290)
* Synchronize steps into PipelineRun annotations

* Upgrade jenkins-zh/jenkins-client

* Refactor stage status structure

Refactor PipelineRun reconciler

* Remove useless dependency replacement

* Add an API to obtain node details for a PipelineRun

* Revert nodedetails API
